20:53PDT 25-Jun-18: Minor changes. Restructuring and refactoring.
